implemented some more api endpoints

+++ Pleisure/UserSession.cs
@@ -10,7 +10,7 @@ namespace Pleisure
 {
 	public class UserSession : SessionBehavior
 	{
-		public int UserID = -1;
+		public long UserID = -1;
 
 		public bool LoggedIn
 		{
